ES-14814 - refactor
