+++ lib/mable_etl.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative 'mable_etl/version'
-require 'mable_etl/configuration'
+require 'mable_etl/configuration/aws_s3'
 require 'pry'
 
 module MableEtl
@@ -10,11 +10,11 @@ module MableEtl
   end
 
   def self.configuration
-    @configuration ||= Configuration.new
+    @configuration ||= Configuration::AwsS3.new
   end
 
   def self.reset
-    @configuration = Configuration.new
+    @configuration = Configuration::AwsS3.new
   end
 
   def self.configure
